refactor: 클래스 명, 패키지 명, 메소드 명 정정
README.md에 패키지 목록 작성

README.md 수정
