+++ src/test/java/lotto/custom/service/LottoDrawingServiceTest/BonusNumberServiceTest.java
@@ -1,4 +1,4 @@
-package lotto.custom.service.LottoDrawingService;
+package lotto.custom.service.LottoDrawingServiceTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -55,7 +55,7 @@ public class BonusNumberServiceTest {
         List<Integer> winningNumbers = Arrays.asList(1, 2, 3, 4, 5, 6);
         assertThatThrownBy(() -> lottoDrawingService.drawBonusNumber(winningNumbers, "1 2"))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(lotto.custom.common.ErrorMessages.SPACE_BETWEEN_NUMBERS);
+                .hasMessage(lotto.custom.common.ErrorMessages.SPACES_BETWEEN_NUMBERS);
     }
 
     @DisplayName("서비스_보너스번호입력_보너스번호와당첨번호가같을때_테스트") // 고칠 것
